backend/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
beanie==1.26.0
motor==3.6.0
pymongo==4.9.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
slowapi==0.1.9
python-multipart==0.0.12

# AI & Media
google-genai>=1.0.0
cloudinary==1.41.0
httpx==0.28.1
